First stab at proof loop w/ integrated induction and coinduction
Instead of having a fixed technique from {coinduction,induction}
for each hipster invocation, try to determine which is relevant
on a per-conjecture basis.

- Induction is relevant to a conjecture if its conclusion mentions a
type t such that t.induct is a defined fact in the context.
- Ditto for coinduction.

If both are deemed relevant, the proof loop tries both induction and
coinduction. If none are relevant, we try neither.